requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
requests==2.31.0
APScheduler==3.10.4
Pillow>=10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests-oauthlib==1.3.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
